마지막

+++ Blade-of-the-Flame/Component/LogicComponents/Monsters/BaseMonster.h
@@ -23,8 +23,8 @@ protected:
 
 	State state_ = MOVE;
 
-	int hp_ = 0;
-	int maxHp_ = 0;
+	float hp_ = 0;
+	float maxHp_ = 0;
 
 	int exp_ = 0;
 	int dmg_ = 0;
